Restructure and update to work with Clojure 1.4.
Since this was one of my first Clojure programs, it contains a few odd
or idiosyncratic ways of doing things that I would not employ today.

But it's a testimony to the stability of the Clojure API that the
changes I had to make to 'modenize' this program were so minimal. And
I'd prefer not to rewrite history, so I'll leave the warts.